refactor(musea): extract props panel usage-code helpers
